requirements.txt
flask==3.0.3
joblib==1.4.2
xgboost==2.0.3
shap==0.45.1
sqlparse==0.5.0
redis==5.0.1
flask-sqlalchemy==3.1.1
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
gunicorn==22.0.0